src/copaw/multi_agent/dashboard/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pyyaml>=6.0
pydantic>=2.0.0